mpz_garble_core
Module encoding

Types for working with binary encoded values.

Binary garbled circuit protocols encode each bit of a value using a pair of labels. One label to represent the bit value 0, and another to represent the bit value 1.

Most computation one would like to perform does not operate on the bit level, rather it typically involves representations such as integers and byte arrays. This module provides convenient types for working with these representations of values in an encoded form.

Free-XOR

The Free-XOR technique stipulates that a global binary offset is used such that the labels for bit value 1 are generated by XORing the label for bit value 0 with the global offset, ie W_1 = W_0 ^ Delta.

Modules

state

Module containing the states of an encoded value.

Structs

ChaChaEncoder

Encodes values using the ChaCha algorithm.

Delta

Global binary offset used by the Free-XOR technique to create label pairs where W_1 = W_0 ^ Delta.

EqualityCheck

A hash value used in dual-execution mode to check equality of two sets of encodings.

Label

Encoded bit label.

Labels

A collection of labels.

Enums

Decoding

Decoding information for an encoded value.

EncodedValue

An encoded value.

EncodingCommitment

A commitment to the encoding of a value.

ValueError

Error related to encoded values.

Traits

Encode

A trait for encoding values.

Encoder

This trait is used to encode values using a global offset (delta).