mpz_garble_core::encoding
Trait Encoder

pub trait Encoder: Send + Sync {
    // Required methods
    fn seed(&self) -> Vec<u8>;
    fn delta(&self) -> Delta;
    fn encode<T: Encode + BinaryLength>(&self, id: u64) -> T::Encoded;
    fn encode_by_type(&self, id: u64, ty: &ValueType) -> EncodedValue<Full>;
}

This trait is used to encode values using a global offset (delta).

Implementations of this trait should be idempotent, meaning that calling encode multiple times with the same id should return the same result. This way every value is assigned a unique encoding which can be regenerated at any time, independent of the order in which values are encoded.

Required Methods

fn seed(&self) -> Vec<u8>

Returns encoder’s rng seed

fn delta(&self) -> Delta

Returns encoder’s global offset

fn encode<T: Encode + BinaryLength>(&self, id: u64) -> T::Encoded

Encodes a type using the provided stream id

id - Unique id of value

fn encode_by_type(&self, id: u64, ty: &ValueType) -> EncodedValue<Full>

Encodes a type using the provided stream id

id - Unique id of value

ty - Type of value

Object Safety

This trait is not object safe.